% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcN.R
\name{calcn}
\alias{calcn}
\title{Calculate New Value of (Specific) Recursive Sequence}
\usage{
calcn(x, n)
}
\arguments{
\item{x}{a vector of the first three numbers in the sequence}

\item{n}{the integer value in the sequence to return}
}
\value{
the value of the nth element in the sequence
}
\description{
For the recursive sequence xn = xn−1 + ((xn−3 − xn−2)/n), returns the
nth value in the sequence given the first three values.
}
\examples{
calcn(c(2, 4, 3), 4) # returns 2.5
\dontrun{calcn(c(2, 4), 4) #returns an error}
}
