% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04pdcwilbourne-package.R
\docType{package}
\name{hw04pdcwilbourne-package}
\alias{hw04pdcwilbourne}
\alias{hw04pdcwilbourne-package}
\title{hw04pdcwilbourne: Defines and Experiments with a Recursive Sequence}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dcwilbourne/hw04pdcwilbourne}
  \item Report bugs at \url{https://github.com/dcwilbourne/hw04pdcwilbourne/issues}
}

}
\author{
\strong{Maintainer}: David Wilbourne \email{dw4757a@american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
