% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrsv_plot.R
\name{rcrsv_plot}
\alias{rcrsv_plot}
\title{Plots Output for a (Specific) Recursive Sequence}
\usage{
rcrsv_plot(rcrsv_df)
}
\arguments{
\item{rcrsv_df}{A tibble with four columns: the first three digits in the sequence, and the
nth value to calculate. The first three columns typically remain the same,
while the last column increases.}
}
\value{
Returns a plot of the output.
}
\description{
For the recursive sequence xn = xn-1 + ((xn-3 - xn-2)/n), plots the output
of the sequence given as input a tibble with information on the sequence's
beginning values and the specific numbers in the sequence to plot.
}
\examples{
my_data <- tibble::tribble(
~x, ~y, ~z, ~n,
2,4,3,3,
2,4,3,4,
2,4,3,5,
2,4,3,6,
2,4,3,7,
2,4,3,8,
2,4,3,9,
2,4,3,10,
2,4,3,12)

rcrsv_plot(my_data) # plots the nth outcomes of the sequence from the
fourth column

my_wrong_data <- tibble::tribble(
~x, ~y, ~z, ~n, ~p,
2,4,3,3,5,
2,4,3,4,5,
2,4,3,5,5,
2,4,3,6,5,
2,4,3,7,5,
2,4,3,8,5,
2,4,3,9,5,
2,4,3,10,5,
2,4,3,12,5)

\dontrun{rcrsv_plot(my_wrong_data) # returns an error}

}
