% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrsv_solve.R
\name{rcrsv_solve}
\alias{rcrsv_solve}
\title{Calculate New Value of (Specific) Recursive Sequence}
\usage{
rcrsv_solve(x, n)
}
\arguments{
\item{x}{A vector of the first three numbers in the sequence}

\item{n}{The integer value in the sequence to return}
}
\value{
Returns the value of the nth element in the sequence
}
\description{
For the recursive sequence xn = xn-1 + ((xn-3 - xn-2)/n), returns the
nth value in the sequence given the first three values.
}
\examples{
rcrsv_solve(c(2, 4, 3), 4) # returns 2.5
\dontrun{rcrsv_solve(c(2, 4), 4) # returns an error}
}
