% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcN2.R
\name{rcrsv_plot}
\alias{rcrsv_plot}
\title{Plots Output Values for Recursive Sequence}
\usage{
rcrsv_plot(rcrsv_df)
}
\arguments{
\item{rcrsv_df}{a tibble first three digits in the sequence and the nth value to calculate}
}
\value{
a plot of the output values for different values of n
}
\description{
Calcn2 plots the output values of a specific recursive sequence over
time given a tibble with information on the sequence's values as input.
}
\examples{
my_data <- tibble::tribble(
~x, ~y, ~z, ~n,
2,4,3,3,
2,4,3,4,
2,4,3,5,
2,4,3,6,
2,4,3,7,
2,4,3,8,
2,4,3,9,
2,4,3,10,
2,4,3,12)

calcn2(my_data)

}
